+++ beetsplug/fetchart.py
@@ -49,6 +49,7 @@ class Candidate:
     CANDIDATE_EXACT = 1
     CANDIDATE_DOWNSCALE = 2
     CANDIDATE_DOWNSIZE = 3
+    CANDIDATE_DEINTERLACE = 4
 
     MATCH_EXACT = 0
     MATCH_FALLBACK = 1
@@ -72,12 +73,13 @@ class Candidate:
         Return `CANDIDATE_DOWNSCALE` if the file must be rescaled.
         Return `CANDIDATE_DOWNSIZE` if the file must be resized, and possibly
             also rescaled.
+        Return `CANDIDATE_DEINTERLACE` if the file must be deinterlaced.
         """
         if not self.path:
             return self.CANDIDATE_BAD
 
         if (not (plugin.enforce_ratio or plugin.minwidth or plugin.maxwidth
-                 or plugin.max_filesize)):
+                 or plugin.max_filesize or plugin.deinterlace)):
             return self.CANDIDATE_EXACT
 
         # get_size returns None if no local imaging backend is available
@@ -144,6 +146,8 @@ class Candidate:
             return self.CANDIDATE_DOWNSCALE
         elif downsize:
             return self.CANDIDATE_DOWNSIZE
+        elif plugin.deinterlace:
+            return self.CANDIDATE_DEINTERLACE
         else:
             return self.CANDIDATE_EXACT
 
@@ -163,6 +167,8 @@ class Candidate:
                 ArtResizer.shared.resize(max(self.size), self.path,
                                          quality=plugin.quality,
                                          max_filesize=plugin.max_filesize)
+        elif self.check == self.CANDIDATE_DEINTERLACE:
+            self.path = ArtResizer.shared.deinterlace(self.path)
 
 
 def _logged_get(log, *args, **kwargs):
@@ -916,6 +922,7 @@ class FetchArtPlugin(plugins.BeetsPlugin, RequestMixin):
             'lastfm_key': None,
             'store_source': False,
             'high_resolution': False,
+            'deinterlace': False,
         })
         self.config['google_key'].redact = True
         self.config['fanarttv_key'].redact = True
@@ -933,6 +940,7 @@ class FetchArtPlugin(plugins.BeetsPlugin, RequestMixin):
                            confuse.String(pattern=self.PAT_PERCENT)]))
         self.margin_px = None
         self.margin_percent = None
+        self.deinterlace = self.config['deinterlace'].get(bool)
         if type(self.enforce_ratio) is str:
             if self.enforce_ratio[-1] == '%':
                 self.margin_percent = float(self.enforce_ratio[:-1]) / 100